Extract tap_passed's logic into PassedTracker class
