+++ lib/tapping_device.rb
@@ -6,6 +6,7 @@ require "tapping_device/output_payload"
 require "tapping_device/trackable"
 require "tapping_device/exceptions"
 require "tapping_device/trackers/initialization_tracker"
+require "tapping_device/trackers/passed_tracker"
 
 class TappingDevice
 
@@ -31,10 +32,6 @@ class TappingDevice
 
   def tap_on!(object)
     track(object, condition: :tap_on?)
-  end
-
-  def tap_passed!(object)
-    track(object, condition: :tap_passed?)
   end
 
   def tap_assoc!(record)
@@ -177,20 +174,6 @@ class TappingDevice
     model_class = object.class
     associations = model_class.reflections
     associations.keys.include?(tp.callee_id.to_s)
-  end
-
-  def tap_passed?(object, tp)
-    # we don't care about calls from the device instance or helper methods
-    return false if is_from_target?(self, tp)
-    return false if tp.defined_class == TappingDevice::Trackable || tp.defined_class == TappingDevice
-
-    collect_arguments(tp).values.any? do |value|
-      # during comparison, Ruby might perform data type conversion like calling `to_sym` on the value
-      # but not every value supports every conversion methods
-      object == value rescue false
-    end
-  rescue
-    false
   end
 
   def get_method_object_from(target, method_name)
